requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
pyarrow==15.0.0
boto3==1.34.34
python-dotenv==1.0.0
jinja2==3.1.3
python-multipart==0.0.6